Graphics:Add error check when create GraphicBuffer
The register buffer fail is filtered out in frameworks, So the app
may use a failed GraphicBuffer that would make the system crash.

When a new GraphicBuffer is created, it need to be registered in
Graphics driver to support multi-process. The GraphicBuffer may
fail in reginstering it. Need error check here, or the GraphicBuffer
would make the system crash.

Change-Id: I8c7688097ee440ab2e10fcf16f21adf512b8eb69
Author: Gang Wang <gang.a.wang@intel.com>
Signed-off-by: Marc Mao <marc.mao@intel.com>
Signed-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 50141, 27600